Added New Tools for changing Course Data
Path Editor will allow for movement of the Path Markers present on each course. Can be imported/exported to 3DS max via the Path Importer maxscript.

Vert Editor will allow for movement of the Verts in the course. Can be imported/exported to 3DS max via the Vert Importer maxscript.

Item Editor will allow for movement of the Item Boxes, Trees or Piranha Plants located in the course. Can be imported/exported to 3DS max via the Path Importer maxscript.

The Course Exporter has a function for exporting the course's model data as sections. The Section Importer 3DS maxscript can import these models, which are separated and marked with the offset to the command in Segment 6 that draws that object.

The Course Exporter tool has several options to export Segment data, either in a compressed or uncompressed format. It can also compress external Segment files to prepare them for reinsertion into the ROM.

The Course Importer will take Segment Data and reinsert it into the ROM, as well as automatically updating the Course Headers to point to the new data. On an Emulator such as Project64, the game can read data that is arbitrarily written to the end of the ROM without purposefully resizing it. Therefore, and to simplify calculating offsets, the Course Importer will append all Course Data on to the end of the ROM. If a segment is not imported from an external file, it will read the segment from it's original location in the ROM and rewrite it on the end with the other course data.

And you can still make Mario go really fast.